chore(ids): move Hid/HidPrefix into ids, re-export from ensemble
Move the content-addressable identity primitive (Hid, HidPrefix,
HidParseError) from ensemble into ids so light bee/identity crates can
mint Hids without dragging the whole mesh (iroh/rustls/etc). ensemble
re-exports ids::{Hid,HidPrefix,HidParseError} for back-compat, so
existing ensemble::Hid call sites keep compiling.

ids gains a hex dep for the wire-form encoding.
