+++ ids/Cargo.toml
@@ -9,6 +9,7 @@ description = "hum-native 256-bit Crockford-base32 identifiers: 48-bit ms timest
 rand = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 sha2 = { workspace = true }
+hex = { workspace = true }
 uuid = { version = "1", features = ["v5"] }
 
 [dev-dependencies]
